Split CMA tests into focused suites
